requirements.txt
mysql-connector-python
python_dotenv